Adds a IndexJobsByNamespace GCS event handler
IndexJobsByNamespace creates a index of all jobs within
a bucket by the work namespace for the jobs. Jobs that have completed are linked from

  gs://BUCKET/index/job-namespace/JOB_NAME/BUILD_NUMBER

with the contents of that file a job result object and a 'link'
metadata attribute pointing to a gs:// path to the source. The
'state' metadata attribute is set to 'success', 'failure', or
'error' if the passed attribute is not set in the finished.json.